pyproject.toml
[tool.poetry]
name = "aelf-gpt"
version = "0.1.0"
description = ""
authors = ["Ademola Balogun <ademolabalogun91@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
llama-index = "^0.10.55"
llama-index-embeddings-huggingface = "^0.2.2"
llama-index-readers-json = "^0.1.5"
llama-index-vector-stores-mongodb = "^0.1.7"
pymongo = {extras = ["srv"], version = "^4.7.2"}
llama-index-llms-gemini = "^0.1.11"
streamlit = "^1.37.0"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
